capek beut gweh

+++ database/migrations/2025_08_30_103959_add_course_id_to_users.php
@@ -11,7 +11,7 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('course_id')->nullable();
+            $table->unsignedBigInteger('course_id')->nullable();
             $table->foreign('course_id')
               ->references('id')
               ->on('courses')
